Scroll bar height fix when watching collection
When watching a collection via slimscoll-watch attribute the height of
the scrollbar would be calculated before DOM had been fully rendered
meaning when you hove over the container the scrollbar height suddenly
jumps. Wrapped the refresh method in a $timeout block to ensure that
execution of it is queued until the DOM render is complete.